Add initial schema prototype and example record.
The schema covers basics, but likely not everything we'd want to represent.
We can evolve it iteratively as we add new records and see what's missing.
The "required" fields are listed just for the purposes of importing the
schema into an editor. (Some editors won't show UI for non-required fields.)